+++ src/options_vecplot.f90
@@ -26,18 +26,21 @@
 !-------------------------------------------------------------------------
 module settings_vecplot
  implicit none
- integer :: npixvec,minpartforarrow,iVecLegendOnPanel
+ integer, parameter :: ivecstyle_streamlines = 0
+ integer, parameter :: ivecstyle_arrows = 1
+ integer, parameter :: ivecstyle_ironfilings = 2
+ integer :: npixvec,minpartforarrow,iVecLegendOnPanel,ivecstyle
  logical :: UseBackgndColorVecplot, iplotpartvec
- logical :: iVecplotLegend,iplotstreamlines,iplotarrowheads
+ logical :: iVecplotLegend,iplotarrowheads
  logical :: iplotsynchrotron,ihidearrowswherenoparts,iallarrowssamelength
- real :: hposlegendvec,vposlegendvec
+ real :: hposlegendvec,vposlegendvec,streamdensity
  real :: rcrit,zcrit,synchrotronspecindex,uthermcutoff
 
  namelist /vectoropts/ npixvec, UseBackgndColorVecplot,iplotpartvec,&
-          iVecplotLegend,hposlegendvec,vposlegendvec,iplotstreamlines, &
+          iVecplotLegend,hposlegendvec,vposlegendvec, &
           iplotarrowheads,iplotsynchrotron,rcrit,zcrit,synchrotronspecindex, &
           uthermcutoff,ihidearrowswherenoparts,minpartforarrow,iallarrowssamelength,&
-          iVecLegendOnPanel
+          iVecLegendOnPanel,ivecstyle,streamdensity
 
 contains
 
@@ -46,14 +49,15 @@ contains
 !---------------------------------------------
 subroutine defaults_set_vecplot
 
- npixvec = 40        ! pixels in x direction on vector plots
+ npixvec = 40        ! pixels in x direction on vector plots (arrow style)
  UseBackgndColorVecplot = .false. ! plot vector plot using black/white
  iplotpartvec = .true.   ! whether to plot particles on vector plot
  iVecplotLegend = .true.
  iVecLegendOnPanel = 0   ! all panels
  hposlegendvec = 0.02
  vposlegendvec = 2.0      ! same default as time legend; inside the viewport
- iplotstreamlines = .false. ! plot stream lines instead of arrows
+ ivecstyle = ivecstyle_streamlines
+ streamdensity = 1.0
  iplotarrowheads = .true.
  iplotsynchrotron = .false.
  zcrit = 2.5 ! kpc
@@ -82,22 +86,18 @@ subroutine submenu_vecplot(ichoose)
  ians = ichoose
  print "(a)",'--------------- vector plot options -------------------'
 
- if (ians <= 0 .or. ians > 7) then
+ if (ians <= 0 .or. ians > 5) then
     print 10,npixvec,print_logical(UseBackgndColorVecplot), &
-             print_logical(iVecplotLegend),print_logical(iplotstreamlines), &
-             print_logical(iplotarrowheads), &
-             print_logical(ihidearrowswherenoparts), &
-             print_logical(iallarrowssamelength)
+             print_logical(iVecplotLegend),ivecstyle, &
+             print_logical(ihidearrowswherenoparts)
 10  format( &
              ' 0) exit ',/, &
              ' 1) change number of pixels                   (',i4,' )',/, &
              ' 2) use background colour for arrows          ( ',a,' )',/, &
              ' 3) vector plot legend settings               ( ',a,' )',/, &
-             ' 4) plot stream/field lines instead of arrows ( ',a,' )',/, &
-             ' 5) turn arrow heads on/off                   ( ',a,' )',/, &
-             ' 6) hide arrows where there are no particles  ( ',a,' )',/, &
-             ' 7) all arrows same length - ie. direction only ( ',a,' )')
-    call prompt('enter option',ians,0,7)
+             ' 4) vector plot style                         (',i4,' )',/, &
+             ' 5) hide arrows where there are no particles  ( ',a,' )')
+    call prompt('enter option',ians,0,5)
  endif
 !
 !--options
@@ -124,43 +124,44 @@ subroutine submenu_vecplot(ichoose)
     endif
 !------------------------------------------------------------------------
  case(4)
-    iplotstreamlines = .not.iplotstreamlines
-    print "(2(a,/))",' Note: the number of stream lines plotted is determined by', &
-                     ' the "change number of contours" option in the r)ender menu'
-    call prompt('use stream lines instead of arrows? ',iplotstreamlines)
-!------------------------------------------------------------------------
- case(5)
-    iplotarrowheads = .not.iplotarrowheads
-    call prompt('plot arrow heads? ',iplotarrowheads)
-    if (ndim==3 .and. .not.iplotarrowheads) then
-       call prompt(' plot synchrotron map? ',iplotsynchrotron)
-       if (iplotsynchrotron) then
-          if (iutherm < 0 .or. iutherm > numplot) then
-             print "(a)",' Warning: cannot use thermal energy cutoff in synchrotron plots'
-             print "(a)",' (could not locate thermal energy in data columns)'
-          endif
-          call prompt(' enter rcrit for cosmic ray electron distribution exp(-r/rcrit -z/zcrit)',rcrit,0.)
-          call prompt(' enter zcrit for cosmic ray electron distribution exp(-r/rcrit -z/zcrit)',zcrit,0.)
-          call prompt(' enter synchrotron spectral index I_nu = nu^-alpha ',synchrotronspecindex,0.)
-          if (iutherm > 0 .and. iutherm <= numplot) then
-             !--set sensible default value for uthermcutoff
-             if (uthermcutoff < -tiny(uthermcutoff)) then
-                uthermcutoff = 0.5*(lim(iutherm,1) + lim(iutherm,2))
+    if (ndim==3) then
+       call prompt('enter vector plot style (0=streamlines, 1=arrows, 2=iron filings)', &
+                   ivecstyle,0,2)
+    else
+       call prompt('enter vector plot style (0=streamlines, 1=arrows)',ivecstyle,0,1)
+    endif
+    if (ivecstyle == ivecstyle_streamlines) then
+       call prompt('enter streamline density',streamdensity,0.05,10.)
+       call prompt('plot arrow heads on streamlines? ',iplotarrowheads)
+    elseif (ivecstyle == ivecstyle_arrows) then
+       call prompt('plot arrow heads? ',iplotarrowheads)
+       if (ndim==3 .and. .not.iplotarrowheads) then
+          call prompt(' plot synchrotron map? ',iplotsynchrotron)
+          if (iplotsynchrotron) then
+             if (iutherm < 0 .or. iutherm > numplot) then
+                print "(a)",' Warning: cannot use thermal energy cutoff in synchrotron plots'
+                print "(a)",' (could not locate thermal energy in data columns)'
              endif
-             call prompt(' enter threshold thermal energy in current units (u < utherm not used) ',uthermcutoff,0.)
+             call prompt(' enter rcrit for cosmic ray electron distribution exp(-r/rcrit -z/zcrit)',rcrit,0.)
+             call prompt(' enter zcrit for cosmic ray electron distribution exp(-r/rcrit -z/zcrit)',zcrit,0.)
+             call prompt(' enter synchrotron spectral index I_nu = nu^-alpha ',synchrotronspecindex,0.)
+             if (iutherm > 0 .and. iutherm <= numplot) then
+                !--set sensible default value for uthermcutoff
+                if (uthermcutoff < -tiny(uthermcutoff)) then
+                   uthermcutoff = 0.5*(lim(iutherm,1) + lim(iutherm,2))
+                endif
+                call prompt(' enter threshold thermal energy in current units (u < utherm not used) ',uthermcutoff,0.)
+             endif
           endif
        endif
+       call prompt('make all arrows same length (ie. only show direction, not magnitude) ?',iallarrowssamelength)
     endif
 !------------------------------------------------------------------------
- case(6)
+ case(5)
     call prompt('hide vector arrows where there are no particles ? ',ihidearrowswherenoparts)
     if (ihidearrowswherenoparts) then
        call prompt(' enter minimum number of particles in pixel cell for arrow to be plotted ',minpartforarrow,1)
     endif
-!------------------------------------------------------------------------
- case(7)
-    iallarrowssamelength = .not.iallarrowssamelength
-    call prompt('make all arrows same length (ie. only show direction, not magnitude) ?',iallarrowssamelength)
  end select
 
  return